requirements.txt
azure-functions
msal
requests
pandas
openpyxl
pyarrow
azure-storage-blob
